+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.2.0
 ### Added
-- Support category filtering per tab.  
+- Support category filtering per tab.
+- Support key frame button for Sequencer.
 
 ## 0.1.0
 ### Added